requirements.txt
dotenv
azure.cognitiveservices.vision.computervision